dispatcher: cast to uint64_t when using time_t value

+++ src/modules/dispatcher/ds_ht.c
@@ -19,6 +19,7 @@
  */
 
 #include <stddef.h>
+#include <stdint.h>
 #include <regex.h>
 
 #include "../../core/mem/shm_mem.h"
@@ -333,8 +334,8 @@ int ds_ht_dbg(ds_ht_t *dsht)
 		while(it) {
 			LM_ERR("\tcell: %.*s\n", it->callid.len, it->callid.s);
 			LM_ERR("\tduid: %.*s\n", it->duid.len, it->duid.s);
-			LM_ERR("\thid: %u expire: %u initexpire: %u\n", it->cellid,
-					(unsigned int)it->expire, (unsigned int)it->initexpire);
+			LM_ERR("\thid: %u expire: %llu initexpire: %llu\n", it->cellid,
+					(uint64_t)it->expire, (uint64_t)it->initexpire);
 			LM_ERR("\tdset:%d\n", it->dset);
 			it = it->next;
 		}
